Move app.js content into PostsPage component.
