+++ instagram-app/src/components/PostContainer/PostContainer.js
@@ -57,20 +57,20 @@ class PostContainer extends Component {
   }
 }
 
-PostContainer.propTypes = {
-  post: PropTypes.shape({
-    username: PropTypes.string.isRequired,
-    thumbnailUrl: PropTypes.string.isRequired,
-    imageUrl: PropTypes.string.isRequired,
-    likes: PropTypes.number,
-    timestamp: PropTypes.string,
-    comments: PropTypes.arrayOf(
-      PropTypes.shape({
-        username: PropTypes.string,
-        text: PropTypes.string
-      })
-    )
-  })
-};
+// PostContainer.propTypes = {
+//   post: PropTypes.shape({
+//     username: PropTypes.string.isRequired,
+//     thumbnailUrl: PropTypes.string.isRequired,
+//     imageUrl: PropTypes.string.isRequired,
+//     likes: PropTypes.number,
+//     timestamp: PropTypes.string,
+//     comments: PropTypes.arrayOf(
+//       PropTypes.shape({
+//         username: PropTypes.string,
+//         text: PropTypes.string
+//       })
+//     )
+//   })
+// };
 
 export default PostContainer;
